refactoring

+++ Entities/Concrete/AccountFinance.cs
@@ -8,6 +8,6 @@ namespace Entities.Concrete
     public class AccountFinance:IEntity
     {
         public int Id { get; set; }
-        public int Cash { get; set; }
+        public decimal Cash { get; set; }
     }
 }
